+++ packages/bsky/src/api/app/bsky/actor/searchActors.ts
@@ -51,7 +51,7 @@ const skeleton = async (inputs: SkeletonFnInput<Context, Params>) => {
   // add hits total
 
   if (ctx.searchAgent) {
-    // @NOTE cursors wont change on appview swap
+    // @NOTE cursors won't change on appview swap
     const { data: res } =
       await ctx.searchAgent.api.app.bsky.unspecced.searchActorsSkeleton({
         q: term,